nav and optimisation

+++ dev/main.py
@@ -7,32 +7,38 @@ from elements.index import Index
 from elements.images import process_all_images
 from elements.tracks import process_all_tracks
 import config
+import pickle
 
-print_time('krawl','first')
-def crawl(path, parent):
-    global website
-    for e in sorted(path.glob('*')):
-        website.append(identify(e, parent))
-        if e.is_dir():
-            crawl(e, website[-1])
+if __name__ == "__main__":
+    def crawl(path, parent):
+        global website
+        for e in sorted(path.glob('*')):
+            website.append(identify(e, parent))
+            if e.is_dir():
+                crawl(e, website[-1])
 
-print_time('metadata','')
-[MetaData(e) for e in Path(config.input).glob('**/*.txt')]
-[MetaData(e) for e in Path(config.input).glob('**/*.tsv')]
+    print_time('metadata','first')
+    [MetaData(e) for e in Path(config.input).glob('**/*.txt')]
+    [MetaData(e) for e in Path(config.input).glob('**/*.tsv')]
 
-print_time('website','')
-website = [Index(Path(config.input))]
-crawl(Path(config.input), website[0])
+    print_time('website','')
+    website = [Index(Path(config.input))]
+    crawl(Path(config.input), website[0])
 
-rmtree(config.output + '/')
+    if Path(config.output).exists():
+        rmtree(config.output + '/')
 
-print_time('process all tracks','')
-process_all_tracks()
+    print_time('tracks','')
+    process_all_tracks()
 
-print_time('process all images','')
-process_all_images()
+    print_time('images','')
+    process_all_images()
 
-print_time('html','')
-website[0].html()
+    print_time('html ','')
+    website[0].html()
 
-print_time('bilou','last')
+    print_time('pickle','')
+    with open('data.pickle', 'wb') as f:
+        pickle.dump(website, f, pickle.HIGHEST_PROTOCOL)
+
+    print_time('bilou','last')
